fix: bug that caused the container to be affected
Previous commit added a Darken feature that gradually darkened the cell
the mouse pointer would hover over by changing the opacity of the said
cell to gradually make it transparent making the black background color
of the cell container div visible.

Since the entire application takes advantage of event delegation by
placing the event listener on the container element itself and using
e.target to change the color of the cells, the other functions would
also change the background color of the container div as well, which
resulted in the background color being a random color instead of black.